requirements.txt
python-telegram-bot[ext]
apscheduler
sqlalchemy
python-dotenv
psycopg2-binary # Driver untuk PostgreSQL